Bug Fix : 좋아요 기능과 모임 퇴장 로직에 현재 모임에 참가하고 있는지 확인 로직 추가

+++ src/main/java/backend/dev/activity/service/ActivityService.java
@@ -128,8 +128,14 @@ public class ActivityService {
     public void activityQuit(Long activityId){
         Activity activity = activityRepository.findById(activityId).orElseThrow(ActivityException.ACTIVITY_NOT_FOUND::getException);
         String userId = SecurityContextHolder.getContext().getAuthentication().getName();
-        userRepository.findById(userId).orElseThrow();
-        activityParticipantsRepository.deleteByUserId(userId);
+        User user = userRepository.findById(userId).orElseThrow(() -> new PublicPlusCustomException(ErrorCode.NOT_FOUND_USER));
+
+        if (activityParticipantsRepository.existsByActivityAndUserId(activity,userId)) {
+            activityParticipantsRepository.deleteByUserId(userId);
+        }else {
+            throw ActivityException.ACTIVITY_NOT_FOUND.getException();
+        }
+
         activity.changeCurrentParticipants(activity.getCurrentParticipants()-1);
 
     }
